packages/training/pyproject.toml
[project]
name = "agent-training-system"
version = "1.0.0"
description = "Periodic training and skill reinforcement for AI agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "chromadb>=0.4.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pynvml>=11.0.0"
]

[project.scripts]
training = "training.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/training"]
